🔨 Move `config.yaml` models to `config.yaml.ts`
